Official File Structure
To avoid having multiple files for things like dbconnections, and multiple img files, I made a simple a file structure.
I envision the files specific only to the admin will be in the admin file.
The files specific only to the customer user side will be only in the custom side.
And anything that is shared, will be in the top level file.  
For instance, there is no reason to have two img directories and two img files.